membuat sistem login

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Post;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -12,6 +13,10 @@ class PostController extends Controller
      */
     public function index()
     {
+        if(!Auth::check()){
+            return redirect('login');
+        } 
+
         $posts = Post::active()->get();
 
         $view_data = [
@@ -25,6 +30,10 @@ class PostController extends Controller
      */
     public function create()
     {
+        if(!Auth::check()){
+            return redirect('login');
+        } 
+
         return view('posts.create');
     }
 
@@ -33,6 +42,10 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        if(!Auth::check()){
+            return redirect('login');
+        } 
+
         $title = $request->input('title');
         $content = $request->input('content');
         
@@ -49,6 +62,10 @@ class PostController extends Controller
      */
     public function show(string $id)
     {
+        if(!Auth::check()){
+            return redirect('login');
+        } 
+
         $post = Post::where('id', '=', $id)->first();
         $comments = $post->comments()->limit(2)->get();
         $total_comments = $post->total_comments();
@@ -67,6 +84,10 @@ class PostController extends Controller
      */
     public function edit(string $id)
     {
+        if(!Auth::check()){
+            return redirect('login');
+        } 
+
         $post = Post::where('id', '=', $id)->first();
 
         $view_data = [
@@ -81,6 +102,10 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if(!Auth::check()){
+            return redirect('login');
+        } 
+
         $title = $request->input('title');
         $content = $request->input('content');
 
@@ -99,6 +124,10 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
+        if(!Auth::check()){
+            return redirect('login');
+        } 
+        
         Post::where('id', $id)->delete();
 
         return redirect('posts');
